+++ src/rfc3339/partial_time/owned.rs
@@ -7,11 +7,13 @@ use core::{convert::TryFrom, fmt, ops, str};
 
 use alloc::{string::String, vec::Vec};
 
-use crate::Error;
+use crate::{ConversionError, Error};
 
 use super::{validate_bytes, PartialTimeStr};
 
 /// Owned string for a time in RFC 3339 [`partial-time`] format, such as `12:34:56.7890`.
+///
+/// Available when `alloc` feature is enabled.
 ///
 /// This is "partial", because it is not associated to a time offset.
 ///
@@ -212,15 +214,32 @@ impl TryFrom<&str> for PartialTimeString {
 }
 
 impl TryFrom<Vec<u8>> for PartialTimeString {
-    type Error = Error;
+    type Error = ConversionError<Vec<u8>>;
 
     #[inline]
     fn try_from(v: Vec<u8>) -> Result<Self, Self::Error> {
-        validate_bytes(&v)?;
-        Ok(unsafe {
-            // This is safe because the value is successfully validated.
-            Self::from_bytes_maybe_unchecked(v)
-        })
+        match validate_bytes(&v) {
+            Ok(_) => Ok(unsafe {
+                // This is safe because the value is successfully validated.
+                Self::from_bytes_maybe_unchecked(v)
+            }),
+            Err(e) => Err(ConversionError::new(v, e)),
+        }
+    }
+}
+
+impl TryFrom<String> for PartialTimeString {
+    type Error = ConversionError<String>;
+
+    #[inline]
+    fn try_from(v: String) -> Result<Self, Self::Error> {
+        match validate_bytes(v.as_bytes()) {
+            Ok(_) => Ok(unsafe {
+                // This is safe because the value is successfully validated.
+                Self::from_bytes_maybe_unchecked(v.into_bytes())
+            }),
+            Err(e) => Err(ConversionError::new(v, e)),
+        }
     }
 }